requirements.txt
backports.shutil-get-terminal-size==1.0.0
decorator==4.4.0
Django==1.11.24
djangorestframework==3.9.4
enum34==1.1.6
ipython==5.8.0
ipython-genutils==0.2.0
pathlib2==2.3.4
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==1.0.16
ptyprocess==0.6.0
Pygments==2.4.2
pytz==2019.2
scandir==1.10.0
simplegeneric==0.8.1
six==1.12.0
traitlets==4.3.2
wcwidth==0.1.7
psycopg2==2.8.3
